refactor moneyvalue for ts
