+++ packages/ui/index.js
@@ -4,7 +4,7 @@ export * from './dist/message.js';
 export * from './dist/numpad-button.js';
 export * from './button-cluster.js';
 export * from './input.js';
-export * from './money-value.js';
+export * from './dist/money-value.js';
 export * from './bitcoin-value.js';
 export * from './dist/option-dot.js';
 export * from './bitcoin-qr-display.js';
