+++ packages/ui/package.json
@@ -54,7 +54,7 @@
     "./numpad-button.js": { "types": "./types/react-jsx.d.ts", "default": "./dist/numpad-button.js" },
     "./button-cluster.js": { "types": "./types/react-jsx.d.ts", "default": "./button-cluster.js" },
     "./input.js": { "types": "./types/react-jsx.d.ts", "default": "./input.js" },
-    "./money-value.js": { "types": "./types/react-jsx.d.ts", "default": "./money-value.js" },
+    "./money-value.js": { "types": "./types/react-jsx.d.ts", "default": "./dist/money-value.js" },
     "./bitcoin-value.js": { "types": "./types/react-jsx.d.ts", "default": "./bitcoin-value.js" },
     "./option-dot.js": { "types": "./types/react-jsx.d.ts", "default": "./dist/option-dot.js" },
     "./bitcoin-qr-display.js": { "types": "./types/react-jsx.d.ts", "default": "./bitcoin-qr-display.js" },
